# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: API BS UTP Parking V1
  description: 'Permite registrar las solicitudes de usuarios para contar con el beneficio de usar los estacionamientos de la UTP.'
  version: 1.0.0
servers:
  - url: /utp-parking/v1
...
paths:
...
  /vehicles/me:
    get:
      operationId: getMyVehicles
      summary: Obtiene los vehículos registrados por el usuario autenticado.
      description: |-
        ### Acerca de la funcionalidad expuesta
        Devuelve los vehículos que el usuario autenticado tiene registrados, habilitados y deshabilitados, junto con la cantidad de vehículos activos y el máximo de activos permitido. El propietario se determina a partir del usuario autenticado, no del cuerpo ni de la ruta de la petición.

        ### Data de Prueba
        No aplica.
      tags:
        - Vehicle Controller
      parameters:
        - $ref: '#/components/parameters/Request-ID'
        - $ref: '#/components/parameters/request-date'
        - $ref: '#/components/parameters/app-code'
        - $ref: '#/components/parameters/caller-name'
      responses:
        '200':
          description: Listado de vehículos del usuario autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDetailList'
        '400':
          description: Los datos proporcionados no son válidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '401':
          description: No está autorizado para consumir el API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '403':
          description: No tiene los permisos necesarios para ejecutar esta operacion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '500':
          description: Ocurrió un error inesperado. Por favor contactarse con Soporte T&eacute;cnico.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '503':
          description: El servicio no se encuentra disponible. Por favor reintente más tarde.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '504':
          description: Se superó el tiempo de respuesta interna.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
  /vehicles:
    post:
      operationId: registerVehicle
      summary: Registra un vehículo para el usuario autenticado.
      description: |-
        ### Acerca de la funcionalidad expuesta
        Registra un vehículo reutilizable que el usuario podrá seleccionar al crear solicitudes de ingreso al estacionamiento. El propietario se determina a partir del usuario autenticado. Un usuario no puede superar el máximo de vehículos activos; los vehículos deshabilitados permanecen registrados y no cuentan para ese límite.

        ### Data de Prueba
        No aplica.
      tags:
        - Vehicle Controller
      parameters:
        - $ref: '#/components/parameters/Request-ID'
        - $ref: '#/components/parameters/request-date'
        - $ref: '#/components/parameters/app-code'
        - $ref: '#/components/parameters/caller-name'
      requestBody:
        description: Entrada del Servicio.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleIn'
        required: true
      responses:
        '201':
          description: Proceso de Registro Satisfactorio.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDetail'
        '400':
          description: Los datos proporcionados no son válidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '401':
          description: No está autorizado para consumir el API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '403':
          description: La placa indicada pertenece a otro usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '409':
          description: El usuario alcanzó el máximo de vehículos activos o ya tiene registrada la placa indicada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '500':
          description: Ocurrió un error inesperado. Por favor contactarse con Soporte T&eacute;cnico.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '503':
          description: El servicio no se encuentra disponible. Por favor reintente más tarde.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '504':
          description: Se superó el tiempo de respuesta interna.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
  /vehicles/{vehicleId}:
    patch:
      operationId: updateVehicleAvailability
      summary: Habilita o deshabilita un vehículo del usuario autenticado.
      description: |-
        ### Acerca de la funcionalidad expuesta
        Modifica únicamente la disponibilidad del vehículo para nuevas solicitudes. Deshabilitar un vehículo no lo elimina ni afecta las solicitudes ni el historial de workflow ya registrados. Habilitar un vehículo requiere que el usuario no haya alcanzado el máximo de vehículos activos. Solo el propietario del vehículo puede ejecutar esta operación.

        ### Data de Prueba
        No aplica.
      tags:
        - Vehicle Controller
      parameters:
        - $ref: '#/components/parameters/Request-ID'
        - $ref: '#/components/parameters/request-date'
        - $ref: '#/components/parameters/app-code'
        - $ref: '#/components/parameters/caller-name'
        - name: vehicleId
          in: path
          description: ID del vehículo.
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        description: Entrada del Servicio.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleAvailabilityIn'
        required: true
      responses:
        '200':
          description: Proceso de Actualización Satisfactorio.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleDetail'
        '400':
          description: Los datos proporcionados no son válidos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '401':
          description: No está autorizado para consumir el API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '403':
          description: El vehículo indicado pertenece a otro usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '404':
          description: El vehículo indicado no existe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '409':
          description: El usuario ya alcanzó el máximo de vehículos activos y no puede habilitar otro.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '500':
          description: Ocurrió un error inesperado. Por favor contactarse con Soporte T&eacute;cnico.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '503':
          description: El servicio no se encuentra disponible. Por favor reintente más tarde.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '504':
          description: Se superó el tiempo de respuesta interna.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
  /vehicles/{vehicleId}/unassignment-requests:
    post:
      operationId: requestVehicleUnassignment
      summary: Solicita la desasignación de un vehículo del usuario autenticado.
      description: |-
        ### Acerca de la funcionalidad expuesta
        Registra una solicitud para que el vehículo deje de pertenecer al usuario. La desasignación no es inmediata: queda sujeta a la revisión de Personal SAE mediante el workflow del proceso VEHICLE_UNASSIGNMENT. Solo el propietario actual puede solicitarla, el vehículo no puede estar ya desasignado y no puede existir otra solicitud abierta para el mismo vehículo.

        ### Data de Prueba
        No aplica.
      tags:
        - Vehicle Controller
      parameters:
        - $ref: '#/components/parameters/Request-ID'
        - $ref: '#/components/parameters/request-date'
        - $ref: '#/components/parameters/app-code'
        - $ref: '#/components/parameters/caller-name'
        - name: vehicleId
          in: path
          description: ID del vehículo.
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        description: Entrada del Servicio.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleUnassignmentIn'
        required: true
      responses:
        '201':
          description: Proceso de Registro Satisfactorio.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleUnassignmentRequestDetail'
        '400':
          description: Los datos proporcionados no son válidos, o falta el motivo de la desasignación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '401':
          description: No está autorizado para consumir el API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '403':
          description: El vehículo indicado pertenece a otro usuario.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '404':
          description: El vehículo indicado no existe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '409':
          description: El vehículo ya está desasignado o ya tiene una solicitud de desasignación en curso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '500':
          description: Ocurrió un error inesperado. Por favor contactarse con Soporte T&eacute;cnico.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '503':
          description: El servicio no se encuentra disponible. Por favor reintente más tarde.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '504':
          description: Se superó el tiempo de respuesta interna.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
  /vehicles/unassignment-requests/me:
    get:
      operationId: getMyVehicleUnassignmentRequests
      summary: Obtiene las solicitudes de desasignación registradas por el usuario autenticado.
      description: |-
        ### Acerca de la funcionalidad expuesta
        Devuelve las solicitudes de desasignación del usuario autenticado con su estado y su historial de workflow, para que pueda conocer cuáles están en revisión y el motivo de las que fueron rechazadas.

        ### Data de Prueba
        No aplica.
      tags:
        - Vehicle Controller
      parameters:
        - $ref: '#/components/parameters/Request-ID'
        - $ref: '#/components/parameters/request-date'
        - $ref: '#/components/parameters/app-code'
        - $ref: '#/components/parameters/caller-name'
      responses:
        '200':
          description: Listado de solicitudes de desasignación del usuario autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleUnassignmentRequestList'
        '401':
          description: No está autorizado para consumir el API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '403':
          description: No tiene los permisos necesarios para ejecutar esta operacion.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '500':
          description: Ocurrió un error inesperado. Por favor contactarse con Soporte T&eacute;cnico.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '503':
          description: El servicio no se encuentra disponible. Por favor reintente más tarde.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '504':
          description: Se superó el tiempo de respuesta interna.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
  /vehicles/unassignment-requests/acceptor/{acceptorId}:
    get:
      operationId: getVehicleUnassignmentRequestsByAcceptor
      summary: Obtiene las solicitudes de desasignación asignadas a un aceptante.
      description: |-
        ### Acerca de la funcionalidad expuesta
        Devuelve las solicitudes de desasignación de vehículos asignadas al Personal SAE indicado, para su revisión. El aceptante consultado debe coincidir con el usuario autenticado.

        ### Data de Prueba
        No aplica.
      tags:
        - Vehicle Controller
      parameters:
        - $ref: '#/components/parameters/Request-ID'
        - $ref: '#/components/parameters/request-date'
        - $ref: '#/components/parameters/app-code'
        - $ref: '#/components/parameters/caller-name'
        - name: acceptorId
          in: path
          description: ID del aceptante.
          required: true
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: Listado de solicitudes de desasignación del aceptante.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleUnassignmentRequestList'
        '401':
          description: No está autorizado para consumir el API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '403':
          description: El usuario autenticado no tiene el rol ROLE_SAE, o no es el aceptante consultado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '500':
          description: Ocurrió un error inesperado. Por favor contactarse con Soporte T&eacute;cnico.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '503':
          description: El servicio no se encuentra disponible. Por favor reintente más tarde.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '504':
          description: Se superó el tiempo de respuesta interna.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
  /vehicles/unassignment-requests/{unassignmentRequestId}:
    patch:
      operationId: respondToVehicleUnassignmentRequest
      summary: Registra la respuesta de Personal SAE a una solicitud de desasignación de vehículo.
      description: |-
        ### Acerca de la funcionalidad expuesta
        Permite que el Personal SAE asignado como aceptante apruebe o rechace la desasignación. Al aprobarla, el vehículo queda en estado UNASSIGNED y sin propietario, conservando intactas sus solicitudes de estacionamiento y su historial. Al rechazarla, el vehículo no se modifica y el motivo del rechazo queda visible para el solicitante.

        ### Data de Prueba
        No aplica.
      tags:
        - Vehicle Controller
      parameters:
        - $ref: '#/components/parameters/Request-ID'
        - $ref: '#/components/parameters/request-date'
        - $ref: '#/components/parameters/app-code'
        - $ref: '#/components/parameters/caller-name'
        - name: unassignmentRequestId
          in: path
          description: ID de la solicitud de desasignación.
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        description: Entrada del Servicio.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParkingResponseIn'
        required: true
      responses:
        '200':
          description: Proceso de Respuesta Satisfactorio.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleUnassignmentRequestDetail'
        '400':
          description: Los datos proporcionados no son válidos, o falta el comentario requerido al rechazar la solicitud.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '401':
          description: No está autorizado para consumir el API
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '403':
          description: El usuario autenticado no tiene el rol ROLE_SAE, o no es el aceptante asignado a la solicitud.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '404':
          description: La solicitud de desasignación indicada no existe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '409':
          description: La solicitud no se encuentra en un estado que permita ser respondida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '500':
          description: Ocurrió un error inesperado. Por favor contactarse con Soporte T&eacute;cnico.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '503':
          description: El servicio no se encuentra disponible. Por favor reintente más tarde.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
        '504':
          description: Se superó el tiempo de respuesta interna.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiException'
...
components:
  schemas:
...
    VehicleInformation:
      type: object
      properties:
        vehicleType:
          type: string
          description: Tipo del vehículo.
          example: Motocicleta
        numberPlate:
          type: string
          description: Placa del vehículo.
          example: HNC-234
    VehicleStatus:
      type: string
      description: >-
        Estado del vehículo dentro de su ciclo de vida. ACTIVE: asignado al usuario y disponible
        para nuevas solicitudes. DISABLED: sigue perteneciendo al usuario pero no puede usarse en
        nuevas solicitudes. UNASSIGNED: ya no pertenece a ningún usuario; el registro se conserva
        para mantener la trazabilidad del historial.
      enum:
        - ACTIVE
        - DISABLED
        - UNASSIGNED
      example: ACTIVE
    VehicleDetail:
      type: object
      description: Vehículo registrado por un usuario. Es reutilizable entre ciclos y entre solicitudes.
      properties:
        idVehicle:
          type: integer
          description: Identificador del vehículo.
          example: 1
        numberPlate:
          type: string
          description: Placa del vehículo.
          example: HNC-234
        idVehicleType:
          type: integer
          description: Identificador del tipo del vehículo.
          example: 2
        vehicleType:
          type: string
          description: Tipo del vehículo.
          example: Motocicleta
        status:
          $ref: '#/components/schemas/VehicleStatus'
    VehicleDetailList:
      type: object
      properties:
        vehicles:
          type: array
          description: >-
            Lista de vehículos asignados al usuario, activos y deshabilitados. Los vehículos
            desasignados no pertenecen a ningún usuario y no se incluyen.
          items:
            $ref: '#/components/schemas/VehicleDetail'
        assignedVehicles:
          type: integer
          description: Cantidad de vehículos asignados al usuario, incluyendo los deshabilitados.
          example: 3
        activeVehicles:
          type: integer
          description: Cantidad de vehículos activos del usuario. Es el valor que se compara contra el máximo permitido.
          example: 2
        maxActiveVehicles:
          type: integer
          description: >-
            Cantidad máxima de vehículos activos que un usuario puede tener. Los vehículos
            deshabilitados permanecen asignados y no cuentan para este máximo; los desasignados
            tampoco, porque ya no pertenecen al usuario.
          example: 5
    VehicleUnassignmentIn:
      type: object
      required:
        - reason
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 255
          description: Motivo por el que el usuario solicita dejar de ser propietario del vehículo.
          example: Vendí el vehículo.
    VehicleUnassignmentRequestDetail:
      type: object
      description: >-
        Solicitud de desasignación de un vehículo. Su aprobación por Personal SAE deja el vehículo
        en estado UNASSIGNED y sin propietario, conservando su historial de solicitudes.
      properties:
        idUnassignmentRequest:
          type: integer
          description: Identificador de la solicitud de desasignación.
          example: 1
        applicant:
          $ref: '#/components/schemas/ApplicantInformation'
        vehicle:
          $ref: '#/components/schemas/VehicleInformation'
        idVehicle:
          type: integer
          description: Identificador del vehículo cuya desasignación se solicita.
          example: 1
        reason:
          type: string
          description: Motivo indicado por el solicitante.
          example: Vendí el vehículo.
        status:
          type: string
          description: Estado de la solicitud de desasignación.
          example: EN_REVISION
        dateRequest:
          type: string
          description: Fecha de la solicitud.
          example: '2026-06-01T10:00:00Z'
        dateResponse:
          type: string
          description: Fecha de la respuesta de Personal SAE, si ya fue atendida.
          example: '2026-06-02T10:00:00Z'
          nullable: true
        workflow:
          type: array
          description: Historial de cambios de estado, del más antiguo al más reciente.
          items:
            $ref: '#/components/schemas/WorkflowEntry'
    VehicleUnassignmentRequestList:
      type: object
      properties:
        unassignmentRequests:
          type: array
          description: Lista de solicitudes de desasignación de vehículos.
          items:
            $ref: '#/components/schemas/VehicleUnassignmentRequestDetail'
    VehicleIn:
      type: object
      required:
        - vehicleType
        - numberPlate
      properties:
        vehicleType:
          type: integer
          description: Tipo del vehículo.
          example: 1
        numberPlate:
          type: string
          description: >-
            Placa del vehículo. Los automóviles y camionetas usan el formato de 3-3 caracteres (ABC-123)
            y las motocicletas, trimotos y motos eléctricas el formato de 2-4 caracteres (AB-1234).
          example: HNC-234
    VehicleAvailabilityIn:
      type: object
      required:
        - active
      properties:
        active:
          type: boolean
          description: Nueva disponibilidad del vehículo para ser utilizado en nuevas solicitudes.
          example: false
    ApplicantInformation:
      type: object
      properties:
        idApplicant:
          type: integer
          description: Identificador del solicitante.
          example: 1
        nameApplicant:
          type: string
          description: Nombre del solicitante.
          example: Juan
        lastNameApplicant:
          type: string
          description: Apellido del solicitante.
          example: Pérez
        usernameApplicant:
          type: string
          description: Nombre de usuario del solicitante.
          example: U23201703
        numberCycle:
          type: string
          description: Número de ciclo de la solicitud.
          example: 2024-1
...
    ParkingResponseIn:
      type: object
      required:
        - approved
      properties:
        approved:
          type: boolean
          description: Indica si la solicitud fue aprobada o rechazada por Personal SAE.
          example: true
        comment:
          type: string
          description: Comentario del Personal SAE sobre la solicitud. Es obligatorio cuando approved es false.
          example: La documentación presentada no es válida.
...
    ApiException:
      title: ApiException
      type: object
      properties:
        description:
          type: string
          description: Descripcion del error de Sistema
          example: Error al llamar al servicio
        errorType:
          type: string
          description: Tipo de Error de Sistema
          example: TECHNICAL
        exceptionDetails:
          type: array
          description: Lista de detalles del error
          items:
            $ref: '#/components/schemas/ApiExceptionDetail'
        properties:
          type: object
          properties: {}
          description: Lista de propiedades especiales.
      description: Datos del error de sistema.
    ApiExceptionDetail:
      title: ApiExceptionDetail
      type: object
      properties:
        component:
          type: string
          description: Nombre del componente de falla
          example: business-parking-request
        description:
          type: string
          description: Descripcion del Detalle
          example: Date is not within the valid range for any cycle
      description: Datos del error técnico.
  parameters:
    Request-ID:
      name: Request-ID
      in: header
      description: Este campo es un valor estandar ya existente y sera usado como identificador.
      required: true
      schema:
        maxLength: 36
        pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[34][0-9a-fA-F]{3}-[89ab][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
        type: string
        example: 550e8400-e29b-41d4-a716-446655440000
    request-date:
      name: request-date
      in: header
      description: Fecha de la petición
      required: true
      schema:
        maxLength: 29
        pattern: '^\d{4}-\d{2}-\d{2}T\d{2}:\d{2}:\d{2}\.\d{3}[+-]\d{4}([a-z]|[A-Z]|\d)?$'
        type: string
        example: '2020-10-09T14:02:03.987-0700'
    app-code:
      name: app-code
      in: header
      description: 'Codigo de la aplicacion que invoca el servicio. Se debe usar el codigo de 2 caracteres que tienen asignada las aplicaciones, puede ser el canal.'
      required: true
      schema:
        maxLength: 2
        minLength: 2
        pattern: '^[0-9A-Z]*$'
        type: string
        example: P0
    caller-name:
      name: caller-name
      in: header
      description: Nombre de la API que realiza la invocacion al servicio.
      required: true
      schema:
        maxLength: 100
        minLength: 5
        pattern: '^[0-9a-zA-Z.-]*$'
        type: string
        example: business-utp-parking-response-v1
  securitySchemes:
    bearerAuth:
      type: http
      description: 'Token de Authorización, header Authorization, ejemplo: Bearer 363536363jjxnj63633611'
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []